+++ oap-server/server-core/src/test/java/org/apache/skywalking/oap/server/core/analysis/meter/function/latest/LatestLabeledFunctionTest.java
@@ -87,37 +87,43 @@ public class LatestLabeledFunctionTest {
 
     @Test
     public void testToHour() {
-        function.accept(MeterEntity.newService("service-test", Layer.GENERAL), HTTP_CODE_COUNT_1);
-        function.accept(MeterEntity.newService("service-test", Layer.GENERAL), HTTP_CODE_COUNT_2);
+        MeterEntity meterEntity1 = MeterEntity.newService("service-test", Layer.GENERAL);
+        meterEntity1.setAttr0("testAttr");
+        function.accept(meterEntity1, HTTP_CODE_COUNT_1);
+        MeterEntity meterEntity2 = MeterEntity.newService("service-test", Layer.GENERAL);
+        meterEntity2.setAttr0("testAttr");
+        function.accept(meterEntity2, HTTP_CODE_COUNT_2);
         function.calculate();
 
         final LatestLabeledFunction hourFunction = (LatestLabeledFunction) function.toHour();
         hourFunction.calculate();
 
         assertThat(hourFunction.getValue()).isEqualTo(HTTP_CODE_COUNT_2);
+        assertThat(hourFunction.getAttr0()).isEqualTo("testAttr");
     }
 
     @Test
     public void testToDay() {
-        function.accept(
-            MeterEntity.newService("service-test", Layer.GENERAL),
-            HTTP_CODE_COUNT_1
-        );
-        function.accept(
-            MeterEntity.newService("service-test", Layer.GENERAL),
-            HTTP_CODE_COUNT_2
-        );
+        MeterEntity meterEntity1 = MeterEntity.newService("service-test", Layer.GENERAL);
+        meterEntity1.setAttr0("testAttr");
+        function.accept(meterEntity1, HTTP_CODE_COUNT_1);
+        MeterEntity meterEntity2 = MeterEntity.newService("service-test", Layer.GENERAL);
+        meterEntity2.setAttr0("testAttr");
+        function.accept(meterEntity2, HTTP_CODE_COUNT_2);
         function.calculate();
 
         final LatestLabeledFunction dayFunction = (LatestLabeledFunction) function.toDay();
         dayFunction.calculate();
 
         assertThat(dayFunction.getValue()).isEqualTo(HTTP_CODE_COUNT_2);
+        assertThat(dayFunction.getAttr0()).isEqualTo("testAttr");
     }
 
     @Test
     public void testSerialize() {
-        function.accept(MeterEntity.newService("service-test", Layer.GENERAL), HTTP_CODE_COUNT_1);
+        MeterEntity meterEntity = MeterEntity.newService("service-test", Layer.GENERAL);
+        meterEntity.setAttr0("testAttr");
+        function.accept(meterEntity, HTTP_CODE_COUNT_1);
 
         LatestLabeledFunction function2 = new LatestLabeledFunctionInst();
         function2.deserialize(function.serialize().build());
@@ -126,11 +132,14 @@ public class LatestLabeledFunctionTest {
         assertThat(function2.getTimeBucket()).isEqualTo(function.getTimeBucket());
         assertThat(function2.getServiceId()).isEqualTo(function.getServiceId());
         assertThat(function2.getValue()).isEqualTo(function.getValue());
+        assertThat(function2.getAttr0()).isEqualTo(function.getAttr0());
     }
 
     @Test
     public void testBuilder() throws IllegalAccessException, InstantiationException {
-        function.accept(MeterEntity.newService("service-test", Layer.GENERAL), HTTP_CODE_COUNT_1);
+        MeterEntity meterEntity = MeterEntity.newService("service-test", Layer.GENERAL);
+        meterEntity.setAttr0("testAttr");
+        function.accept(meterEntity, HTTP_CODE_COUNT_1);
         function.calculate();
 
         StorageBuilder<LatestLabeledFunction> storageBuilder = function.builder().newInstance();
@@ -143,6 +152,7 @@ public class LatestLabeledFunctionTest {
         LatestLabeledFunction function2 = storageBuilder.storage2Entity(new HashMapConverter.ToEntity(map));
 
         assertThat(function2.getValue()).isEqualTo(function.getValue());
+        assertThat(function2.getAttr0()).isEqualTo(function.getAttr0());
     }
 
     private static class LatestLabeledFunctionInst extends LatestLabeledFunction {